requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
psycopg==3.2.1
qdrant-client==1.11.3
anthropic>=0.49.0
openai>=1.12.0
tiktoken>=0.6.0
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
httpx>=0.27.0
